@@ -42,11 +42,27 @@ private:
     TypeRegister() = default;
 
 public:
-    static TypeRegister<T> New(const std::string& name = empty_string, const std::map<size_t, std::any>& attributes = {})
+    static TypeRegister<T> New(const std::map<size_t, std::any>& attributes = {})
     {
         Type* type = type_of<T>();
-        if (!name.empty())
-            type->m_name = name;
+        type->m_attributes = attributes;
+
+        TypeRegister<T> reg;
+
+        // if constexpr(!is_object<T>)
+        //{
+        //     if constexpr(std::)
+        //     type->m_constructors.push_back(ConstructorInfo::Register<T>(&ctor<T>));
+        //     type->m_constructors.push_back(ConstructorInfo::Register<T>(&ctor<T, const T&>));
+        // }
+
+        return reg;
+    }
+
+    static TypeRegister<T> New(const std::string& name, const std::map<size_t, std::any>& attributes = {})
+    {
+        Type* type = type_of<T>();
+        type->m_name = name;
         type->m_attributes = attributes;
 
         TypeRegister<T> reg;
@@ -71,9 +87,9 @@ public:
     template <typename U>
     TypeRegister<T>& convert()
     {
-        type_of<T>()->m_typeConvertors.push_back({type_of<U>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
+        type_of<T>()->m_typeConvertors.push_back({type_of<U>(), [](const ObjectPtr& obj, Type* targetType, ObjectPtr& target) -> bool
                                                   {
-                                                      using TU = typename rtti::TypeWarper<U>::type;
+                                                      using TU = type_t<U>;
                                                       if constexpr (is_object<T>)
                                                       {
                                                           std::shared_ptr<T> tobj = cast<std::shared_ptr<T>>(obj);
@@ -122,7 +138,8 @@ public:
     }
 
     // R / R& / const R&
-    template <typename R, typename U = std::remove_pointer_t<T>, std::enable_if_t<!std::is_arithmetic_v<U>, bool> = true>
+    template <typename R, typename U = std::remove_pointer_t<T>>
+        requires(!std::is_arithmetic_v<U>)
     TypeRegister<T>& property(const std::string& name, R (U::*getter)(), void (U::*setter)(R), const std::map<size_t, std::any>& attributes = {})
     {
         type_of<T>()->m_properties.push_back(new PropertyInfo(type_of<T>(), name, type_of<R>(), MethodInfo::Register<T, decltype(getter), R>(name, getter), MethodInfo::Register<T, decltype(setter), void, R>(name, setter), attributes));
@@ -130,7 +147,8 @@ public:
     }
 
     // R / R& / const R&
-    template <typename R, typename U = std::remove_pointer_t<T>, std::enable_if_t<!std::is_arithmetic_v<U>, bool> = true>
+    template <typename R, typename U = std::remove_pointer_t<T>>
+        requires(!std::is_arithmetic_v<U>)
     TypeRegister<T>& property(const std::string& name, R (U::*getter)(), const std::map<size_t, std::any>& attributes = {})
     {
         type_of<T>()->m_properties.push_back(new PropertyInfo(type_of<T>(), name, type_of<R>(), MethodInfo::Register<T, decltype(getter), R>(name, getter), nullptr, attributes));
@@ -138,7 +156,8 @@ public:
     }
 
     // R / R& / const R&
-    template <typename R, typename U = std::remove_pointer_t<T>, std::enable_if_t<!std::is_arithmetic_v<U>, bool> = true>
+    template <typename R, typename U = std::remove_pointer_t<T>>
+        requires(!std::is_arithmetic_v<U>)
     TypeRegister<T>& property(const std::string& name, R (U::*getter)() const, void (U::*setter)(R), const std::map<size_t, std::any>& attributes = {})
     {
         type_of<T>()->m_properties.push_back(new PropertyInfo(type_of<T>(), name, type_of<R>(), MethodInfo::Register<T, decltype(getter), R>(name, getter), MethodInfo::Register<T, decltype(setter), void, R>(name, setter), attributes));
@@ -146,7 +165,8 @@ public:
     }
 
     // R / R& / const R&
-    template <typename R, typename U = std::remove_pointer_t<T>, std::enable_if_t<!std::is_arithmetic_v<U>, bool> = true>
+    template <typename R, typename U = std::remove_pointer_t<T>>
+        requires(!std::is_arithmetic_v<U>)
     TypeRegister<T>& property(const std::string& name, R (U::*getter)() const, const std::map<size_t, std::any>& attributes = {})
     {
         type_of<T>()->m_properties.push_back(new PropertyInfo(type_of<T>(), name, type_of<R>(), MethodInfo::Register<T, decltype(getter), R>(name, getter), nullptr, attributes));
@@ -154,7 +174,8 @@ public:
     }
 
     // field
-    template <typename R, bool READONLY = false, typename U = std::remove_pointer_t<T>, std::enable_if_t<!std::is_arithmetic_v<U>, bool> = true>
+    template <typename R, bool READONLY = false, typename U = std::remove_pointer_t<T>>
+        requires(!std::is_arithmetic_v<U>)
     TypeRegister<T>& property(const std::string& name, R U::*field, const std::map<size_t, std::any>& attributes = {})
     {
         MethodInfo* getter = MethodInfo::Register<T, decltype(field), R>(name, field);
@@ -169,14 +190,16 @@ public:
         return *this;
     }
 
-    template <typename R, typename... Args, typename U = std::remove_pointer_t<T>, std::enable_if_t<!std::is_arithmetic_v<U>, bool> = true>
+    template <typename R, typename... Args, typename U = std::remove_pointer_t<T>>
+        requires(!std::is_arithmetic_v<U>)
     TypeRegister<T>& method(const std::string& name, R (U::*func)(Args...), const std::map<size_t, std::any>& attributes = {})
     {
         type_of<T>()->m_methods.push_back(MethodInfo::Register<T, decltype(func), R, Args...>(name, func, attributes));
         return *this;
     }
 
-    template <typename R, typename... Args, typename U = std::remove_pointer_t<T>, std::enable_if_t<!std::is_arithmetic_v<U>, bool> = true>
+    template <typename R, typename... Args, typename U = std::remove_pointer_t<T>>
+        requires(!std::is_arithmetic_v<U>)
     TypeRegister<T>& method(const std::string& name, R (U::*func)(Args...) const, const std::map<size_t, std::any>& attributes = {})
     {
         type_of<T>()->m_methods.push_back(MethodInfo::Register<T, decltype(func), R, Args...>(name, func, attributes));
@@ -199,11 +222,21 @@ private:
     TypeRegister() = default;
 
 public:
-    static TypeRegister<T> New(const std::string& name = empty_string, const std::map<size_t, std::any>& attributes = {})
+    static TypeRegister<T> New(const std::map<size_t, std::any>& attributes = {})
     {
         Type* type = type_of<T>();
-        if (!name.empty())
-            type->m_name = name;
+        type->m_attributes = attributes;
+
+        TypeRegister<T> reg;
+        type->m_constructors.push_back(ConstructorInfo::Register<T>(&ctor<T, const T&>));
+
+        return reg;
+    }
+
+    static TypeRegister<T> New(const std::string& name, const std::map<size_t, std::any>& attributes = {})
+    {
+        Type* type = type_of<T>();
+        type->m_name = name;
         type->m_attributes = attributes;
 
         TypeRegister<T> reg;
@@ -229,83 +262,22 @@ public:
 template <typename T>
 inline Type* DefaultEnumRegister(Type* type)
 {
-    type->m_underlyingType = type_of<typename std::underlying_type<T>::type>();
+    if (!type->m_constructors.empty())
+        return type;
+
+    using UT = std::underlying_type<T>::type;
 
     type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T>));
-    // FIXME: This will trigger infinite recursion here.
-    // type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, const T&>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, int8_t>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, uint8_t>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, int16_t>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, uint16_t>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, int32_t>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, uint32_t>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, int64_t>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, uint64_t>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, long>));
-    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, unsigned long>));
+    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, const T&>));
+    type->m_constructors.push_back(ConstructorInfo::Register(type, &ctor<T, UT>));
 
-    type->m_typeConvertors.push_back({type_of<int8_t>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
+    type->m_typeConvertors.push_back({type_of<UT>(), [](const ObjectPtr& obj, Type* targetType, ObjectPtr& target) -> bool
                                       {
-                                          target = rtti::Box(static_cast<int8_t>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<uint8_t>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<uint8_t>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<int16_t>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<int16_t>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<uint16_t>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<uint16_t>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<int32_t>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<int32_t>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<uint32_t>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<uint32_t>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<int64_t>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<int64_t>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<uint64_t>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<uint64_t>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<long>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<long>(rtti::Unbox<T>(obj)));
-                                          return true;
-                                      }});
-    type->m_typeConvertors.push_back({type_of<unsigned long>(), [](const ObjectPtr& obj, [[maybe_unused]] Type* targetType, ObjectPtr& target) -> bool
-                                      {
-                                          target = rtti::Box(static_cast<unsigned long>(rtti::Unbox<T>(obj)));
+                                          target = rtti::Box(static_cast<UT>(rtti::Unbox<T>(obj)));
                                           return true;
                                       }});
 
-    type->m_objectComparers.push_back({type_of<int8_t>(), &comparer<typename std::underlying_type<T>::type, int8_t>});
-    type->m_objectComparers.push_back({type_of<uint8_t>(), &comparer<typename std::underlying_type<T>::type, uint8_t>});
-    type->m_objectComparers.push_back({type_of<int16_t>(), &comparer<typename std::underlying_type<T>::type, int16_t>});
-    type->m_objectComparers.push_back({type_of<uint16_t>(), &comparer<typename std::underlying_type<T>::type, uint16_t>});
-    type->m_objectComparers.push_back({type_of<int32_t>(), &comparer<typename std::underlying_type<T>::type, int32_t>});
-    type->m_objectComparers.push_back({type_of<uint32_t>(), &comparer<typename std::underlying_type<T>::type, uint32_t>});
-    type->m_objectComparers.push_back({type_of<int64_t>(), &comparer<typename std::underlying_type<T>::type, int64_t>});
-    type->m_objectComparers.push_back({type_of<uint64_t>(), &comparer<typename std::underlying_type<T>::type, uint64_t>});
-    type->m_objectComparers.push_back({type_of<long>(), &comparer<typename std::underlying_type<T>::type, long>});
-    type->m_objectComparers.push_back({type_of<unsigned long>(), &comparer<typename std::underlying_type<T>::type, unsigned long>});
+    type->m_objectComparers.push_back({type_of<T>(), &comparer<T, T>});
 
     return type;
 }
